- pin jax to 0.4.35 - add Mv normal
using this as an opportunity to roll up initial development; further work will be in feature PRs.

JAX changes its method of discovering leaked tracers in a way that exposes a problem with my technique for transforming JAXPRs. The next task is to shift the technique to use the JAX convenience functions for working with JAXPRs that avoid these problems; the way I have done it is "too manual."