moves concrete implementation of repeat into interpreter
this hasn't quite accomplished all we hoped, preventing vmap from dissolving all the interface boundaries we're trying to maintain, but does allow repeat to show up in certain jaxprs but it doesn't persist across the simulate transformation.

Custom batching might allow us to get out of this cul-de-sac, in particular the sequential_vmap decorator. Theory: inside a given shader, we cannot vectorize operations further (from the point of view of the source code), so it makes sense that vmap can be sequential inside any Gen model. What is vectorized is the simulate (importance, assess, ...) operation of the whole model: what we are aiming for is for the top level vmap over keys (to generate rich samples, importance samples or to assess a large number of model variants) will be the point at which we exploit the GPU's parallelism, and not at interior points, which I don't know how to do